update
Using Viewer (1st page), sending a QUERY to the FW to get a RESP of the Pressure Value so it can be rendered in the QML (Circular guage and all) in "real time".

I'm using the query - response method with Viewer sending the QUERY and FW generating the RESP of most recent Pressure Value.